i586: fix JSVALUE32_64 getOwn sentinel and encodeInto size_t truncation
JSVALUE32_64 getOwn sentinel:
JSC__JSValue__getOwn returns EncodedJSValue(0) (= all-bits-zero) when a
property is not found. On JSVALUE64 this matches ValueEmpty (0) and
is_empty() returns true. On JSVALUE32_64, EncodedJSValue(0) decodes to
tag=0/payload=0 -- not EmptyValueTag (0xfffffff9), so is_empty() misses
it. The Rust side thinks the property exists as Int32(0), and
get_own_object throws "files must be an object" for any missing property
(e.g. Bun.build config without a 'files' field).

Fix: check v.0 == 0 in addition to v.is_empty() in get_own(). Catches
the EncodedJSValue(0) sentinel on both architectures.

TextEncoder.encodeInto:
Rust returns a u64 packing (written << 32) | read. The C++ side
declared the return type as size_t (32-bit on i586), truncating the
written field to 0. Changed to uint64_t.

+++ src/jsc/bindings/webcore/JSTextEncoder.cpp
@@ -70,8 +70,8 @@ using namespace JSC::DOMJIT;
 
 extern "C" JSC::EncodedJSValue TextEncoder__encode8(JSC::JSGlobalObject* global, const Latin1Character* stringPtr, size_t stringLen);
 extern "C" JSC::EncodedJSValue TextEncoder__encode16(JSC::JSGlobalObject* global, const char16_t* stringPtr, size_t stringLen);
-extern "C" size_t TextEncoder__encodeInto8(const Latin1Character* stringPtr, size_t stringLen, void* ptr, size_t len);
-extern "C" size_t TextEncoder__encodeInto16(const char16_t* stringPtr, size_t stringLen, void* ptr, size_t len);
+extern "C" uint64_t TextEncoder__encodeInto8(const Latin1Character* stringPtr, size_t stringLen, void* ptr, size_t len);
+extern "C" uint64_t TextEncoder__encodeInto16(const char16_t* stringPtr, size_t stringLen, void* ptr, size_t len);
 extern "C" JSC::EncodedJSValue TextEncoder__encodeRopeString(JSC::JSGlobalObject* lexicalGlobalObject, JSC::JSString* str);
 
 // extern "C" {
@@ -439,7 +439,7 @@ static inline JSC::EncodedJSValue jsTextEncoderPrototypeFunction_encodeIntoBody(
         return {};
     }
 
-    size_t res = 0;
+    uint64_t res = 0;
     if (!source->is8Bit()) {
         const auto span = source->span16();
         res = TextEncoder__encodeInto16(span.data(), span.size(), destination->vector(), destination->byteLength());
